pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fleet-ai"
version = "2.0.0"
description = "Scalable oversight environment for training AI inspector agents that monitor customer support workers."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
train = [
    "unsloth",
    "trl",
    "datasets",
    "transformers",
    "wandb",
]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["fleet_ai", "fleet_ai.server", "fleet_ai.environment"]
package-dir = { "fleet_ai" = ".", "fleet_ai.server" = "server", "fleet_ai.environment" = "environment" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
